requirements.txt
# Core Dependencies for Spacecraft Safe RL Framework
# ===================================================

# Deep Learning and RL Framework
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Reinforcement Learning
gymnasium>=0.29.0
stable-baselines3>=2.0.0
ray[rllib]>=2.7.0

# Scientific Computing and Numerical Analysis
numpy>=1.24.0
scipy>=1.11.0
pandas>=2.0.0

# Orbital Mechanics and Aerospace
# Note: These would be the ideal packages for aerospace applications
# If not available, fallback implementations are provided in the codebase
# orekit>=11.3.0  # Java-based orbital mechanics (requires JPype1)
# poliastro>=0.17.0  # Pure Python alternative
# astropy>=5.3.0  # Astronomy and astrophysics

# Control Systems and Optimization
control>=0.9.4
cvxpy>=1.3.0
casadi>=3.6.0

# Visualization and Analysis
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
dash>=2.11.0

# Data Processing and Statistics
scikit-learn>=1.3.0
statsmodels>=0.14.0

# Configuration and Utilities
pyyaml>=6.0
click>=8.1.0
tqdm>=4.65.0
colorama>=0.4.6

# Logging and Monitoring
tensorboard>=2.13.0
wandb>=0.15.0
mlflow>=2.5.0

# Testing and Development
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0

# Optional GPU acceleration
# Uncomment based on your CUDA version
# torch-geometric>=2.3.0
# cupy-cuda11x>=12.0.0  # For CUDA 11.x
# cupy-cuda12x>=12.0.0  # For CUDA 12.x

# Web Interface (for Streamlit demo)
streamlit>=1.25.0
streamlit-plotly>=0.0.9

# Jupyter Notebooks (for development and examples)
jupyter>=1.0.0
ipywidgets>=8.0.0

# Documentation (optional)
sphinx>=7.0.0
sphinx-rtd-theme>=1.3.0

# Performance Profiling (optional)
line-profiler>=4.0.0
memory-profiler>=0.61.0

# Additional aerospace-specific packages (install if available)
# sgp4>=2.22  # Satellite orbit propagation
# pyephem>=4.1.4  # Astronomical calculations
# skyfield>=1.46  # High precision astronomy